refactor(cli): co-locate command registration in each command folder
Move Commander wiring out of the monolithic cli-program.ts into per-command
registerX(program) functions co-located with each command. createProgram()
now just calls them in registration order, shrinking cli-program.ts from
1109 to ~320 lines and making each command's surface owned by its own folder.

- Each command folder exposes registerX(program): void (mirrors the existing
  registerExtras precedent)
- New commands/toggles/ owns the shared enable/disable parents that wire the
  orgs + billing handlers (neither feature folder cleanly owns them)
- Shared option parsers (parseIntegerOption, collectOptionValues) extracted to
  lib/option-parsers.ts
- clerk --help output verified byte-identical across all 39 command nodes;
  no user-facing change
