+++ packages/cli-core/src/commands/completion/index.ts
@@ -1,3 +1,4 @@
+import { createArgument, type Command } from "@commander-js/extra-typings";
 import { generate as generateBash } from "./shells/bash.ts";
 import { generate as generateZsh } from "./shells/zsh.ts";
 import { generate as generateFish } from "./shells/fish.ts";
@@ -62,4 +63,48 @@ Run 'clerk completion --help' for full setup instructions.`,
   }
   process.stdout.write(GENERATORS[shell]("clerk"));
   printNextSteps(INSTALL_HINTS[shell]);
+}
+
+export function registerCompletion(program: Command): void {
+  program
+    .command("completion")
+    .description("Generate shell autocompletion script")
+    .addArgument(
+      createArgument("[shell]", `Shell type (${SUPPORTED_SHELLS.join(", ")})`).choices(
+        SUPPORTED_SHELLS,
+      ),
+    )
+    .setExamples([
+      { command: "clerk completion bash", description: "Output bash completion script" },
+      { command: "clerk completion zsh", description: "Output zsh completion script" },
+      { command: "clerk completion fish", description: "Output fish completion script" },
+      {
+        command: "clerk completion powershell",
+        description: "Output PowerShell completion script",
+      },
+    ])
+    .addHelpText(
+      "after",
+      `
+Tutorial — enable completions for your shell:
+
+  Bash:
+    $ eval "$(clerk completion bash)"                          # Current session only
+    $ clerk completion bash > /etc/bash_completion.d/clerk     # Permanent (Linux)
+    $ echo 'eval "$(clerk completion bash)"' >> ~/.bashrc      # Permanent (append)
+
+  Zsh:
+    $ eval "$(clerk completion zsh)"                           # Current session only
+    $ mkdir -p ~/.zfunc && clerk completion zsh > ~/.zfunc/_clerk  # Permanent
+    # Then add to ~/.zshrc: fpath=(~/.zfunc $fpath); autoload -Uz compinit && compinit
+
+  Fish:
+    $ mkdir -p ~/.config/fish/completions
+    $ clerk completion fish > ~/.config/fish/completions/clerk.fish  # Auto-discovered
+
+  PowerShell:
+    $ clerk completion powershell | Out-String | Invoke-Expression  # Current session
+    $ clerk completion powershell >> $PROFILE                       # Permanent`,
+    )
+    .action(completion);
 }
