+++ packages/cli-core/src/commands/update/index.ts
@@ -1,3 +1,4 @@
+import type { Command } from "@commander-js/extra-typings";
 import { isAgent, isHuman } from "../../mode.ts";
 import { green, cyan, yellow, dim } from "../../lib/color.ts";
 import { CliError } from "../../lib/errors.ts";
@@ -428,4 +429,23 @@ export async function update(options: UpdateOptions): Promise<void> {
   if (isHuman()) {
     outro(anyFailed ? "Update completed with errors" : `Successfully updated to ${latest}`);
   }
+}
+
+export function registerUpdate(program: Command): void {
+  program
+    .command("update")
+    .description("Update the Clerk CLI to the latest version")
+    .option("--channel <tag>", "Release channel to update to (e.g. latest, canary)")
+    .option("-y, --yes", "Skip confirmation prompt")
+    .option("--all", "Update every clerk install found on PATH, not just the first")
+    .setExamples([
+      { command: "clerk update", description: "Update to the latest stable release" },
+      {
+        command: "clerk update --channel canary",
+        description: "Update to the latest canary release",
+      },
+      { command: "clerk update --yes", description: "Update without confirmation prompt" },
+      { command: "clerk update --all", description: "Update every clerk install on PATH" },
+    ])
+    .action(update);
 }
